+++ doc/changelog.md
@@ -7,6 +7,8 @@
 - Document settings, like paper size, margins, spacings are now stored in saved documents. When loading a document,
   the stored settings overwrite the default values set in the application settings.
   - The Edit menu in the main window has a new option to edit these document settings for the current document only.
+- Added new card download filter for digital cards. The new filter includes both digital “reprints” of existing cards
+  and digital-only cards that aren’t available as physical cards at all.
 
 ## Changed features